customer-api-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-aiplatform==1.70.0
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.20.0
pydantic-settings==2.5.2
httpx==0.27.2
